requirements.txt
openai>=1.47.0
python-dotenv>=1.0.1
python-docx>=1.1.2
pandas>=2.2.2
numpy>=2.0.2
scikit-learn>=1.5.2
sentence-transformers>=3.2.0
faiss-cpu>=1.8.0
matplotlib>=3.9.2
streamlit>=1.38.0
